refactor: simplify generated python code (#558)
(cherry picked from commit ea0331ef5a6d6e9d60fd36291c8f1b4f61dc8abe)
