+++ packages/create-llama/templates/components/loaders/python/file/loader.py
@@ -1,5 +1,6 @@
 from llama_index.core.readers import SimpleDirectoryReader
-from app.engine.constants import DATA_DIR
+
+DATA_DIR = "data"  # directory to cache the generated index
 
 
 def get_documents():
