--- packages/create-llama/templates/components/vectordbs/python/mongo/constants.py
@@ -1,3 +1,0 @@
-DATA_DIR = "data"  # directory containing the documents to index
-CHUNK_SIZE = 1024
-CHUNK_OVERLAP = 20
