+++ packages/create-llama/templates/components/vectordbs/python/none/constants.py
@@ -1,4 +1,1 @@
 STORAGE_DIR = "storage"  # directory to cache the generated index
-DATA_DIR = "data"  # directory containing the documents to index
-CHUNK_SIZE = 1024
-CHUNK_OVERLAP = 20
