+++ packages/create-llama/templates/components/vectordbs/python/none/index.py
@@ -2,7 +2,6 @@ import logging
 import os
 
 from app.engine.constants import STORAGE_DIR
-from app.engine.settings import init_settings
 from llama_index.core.storage import StorageContext
 from llama_index.core.indices import load_index_from_storage
 
@@ -10,9 +9,6 @@ logger = logging.getLogger("uvicorn")
 
 
 def get_index():
-    # Init default app global settings
-    init_settings()
-
     # check if storage already exists
     if not os.path.exists(STORAGE_DIR):
         raise Exception(
