+++ packages/create-llama/templates/components/vectordbs/python/pg/generate.py
@@ -6,9 +6,8 @@ import logging
 from llama_index.core.indices import VectorStoreIndex
 from llama_index.core.storage import StorageContext
 
-from app.engine.constants import DATA_DIR
 from app.engine.loader import get_documents
-from app.engine.settings import init_settings
+from app.settings import init_settings
 from app.engine.utils import init_pg_vector_store_from_env
 
 logging.basicConfig(level=logging.INFO)
