+++ packages/create-llama/templates/components/vectordbs/python/pg/index.py
@@ -1,15 +1,11 @@
 import logging
 from llama_index.core.indices.vector_store import VectorStoreIndex
-from app.engine.settings import init_settings
 from app.engine.utils import init_pg_vector_store_from_env
 
 logger = logging.getLogger("uvicorn")
 
 
 def get_index():
-    # Init default app global settings
-    init_settings()
-
     logger.info("Connecting to index from PGVector...")
     store = init_pg_vector_store_from_env()
     index = VectorStoreIndex.from_vector_store(store)
